Refactor index.js and entry.js and convert to ESM style modules.
